requirements.txt
# Core dependencies
gpt4all>=2.6.0
langchain>=0.3.0
langchain-community>=0.3.0
langchain-core>=0.3.0
langgraph>=0.2.0

# Vector search (RAG over parking_info.txt)
faiss-cpu>=1.7.4

# Vector database (for future Weaviate integration)
weaviate-client>=4.0.0

# Embeddings
sentence-transformers>=2.7.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.9.0
pydantic-settings>=2.5.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=5.0.0

# Evaluation metrics
numpy>=1.26.0,<3.0.0
scikit-learn>=1.5.0

# Guard rails / NLP
torch>=2.4.0
transformers>=4.40.0

# Async support
aiohttp>=3.10.0

# Logging
loguru>=0.7.2

# Linting (optional: pip install ruff)
ruff>=0.4.0
